Stu_com APIs/requirements.txt
Flask
boto3
python-dotenv
bcrypt
pyjwt
